chore: moved testings utils
