+++ packages/fuel-gauge/src/abi/utils.ts
@@ -1,0 +1,19 @@
+import { getForcProject } from '@fuel-ts/utils/test-utils';
+import { join } from 'path';
+
+export enum AbiProjectsEnum {
+  ABI_CONTRACT = 'abi-contract',
+  ABI_PREDICATE = 'abi-predicate',
+  ABI_SCRIPT = 'abi-script',
+}
+
+const forcProjectsDir = join(__dirname, '../../test/fixtures/forc-projects');
+
+export const getAbiForcProject = (project: AbiProjectsEnum) => {
+  const result = getForcProject({
+    projectDir: join(forcProjectsDir, project),
+    projectName: project,
+    build: 'release',
+  });
+  return result;
+};
